chore(deps): break ForceRepository ↔ operations import cycle
Extract ForceErrorCode + IForceOperationResult into a leaf
ForceRepository.types.ts module so the repository class and its operation
helpers no longer import from each other. ForceRepository.ts re-exports
both names so external consumers (api/forces, EncounterService, tests) keep
their existing import paths working.

Cycle removed: forces/ForceRepository.ts ↔ forces/ForceRepository.operations.ts
(madge --circular: 12 → 11)

No behavior change.
